requirements.txt
aiofiles==24.1.0
influxdb_client==1.40.0
python-kasa==0.7.0.1
